+++ app/assets/javascripts/infrastructures.js
@@ -56,53 +56,7 @@
   Vue.component('insert-cf-params',   require('infrastructures/insert-cf-params.js'));
   Vue.component('add-ec2-tabpane',    require('infrastructures/add-ec2-tabpane.js'));
   Vue.component('cf-history-tabpane', require('infrastructures/cf-history-tabpane.js'));
-
-  Vue.component("infra-logs-tabpane", {
-    template: '#infra-logs-tabpane-template',
-
-    props: {
-      infra_id: {
-        type: Number,
-        required: true,
-      },
-    },
-
-    data: function () {return {
-      logs: [],
-      page: {},
-    };},
-
-    methods: {
-      status_class: function (status) { return status ? 'label-success' : 'label-danger'; },
-      status_text: function (status)  { return status ? 'SUCCESS' : 'FAILED'; },
-      toLocaleString: toLocaleString,
-    },
-
-    created: function () {
-      var self = this;
-      console.log(self);
-      var infra = new Infrastructure(this.infra_id);
-      infra.logs().done(function (data) {
-        self.logs = data.logs;
-        self.page = data.page;
-        self.$parent.loading = false;
-      }).fail(alert_and_show_infra);
-
-      this.$watch('infra_logs', function (newVal, oldVal) {
-        $(".popovermore").popover().click( function(e) {
-          e.preventDefault();
-        });
-      });
-
-      this.$on('show', function (page) {
-        var infra = new Infrastructure(self.infra_id);
-        infra.logs(page).done(function (data) {
-          self.logs = data.logs;
-          self.page = data.page;
-        }).fail(alert_and_show_infra);
-      });
-    },
-  });
+  Vue.component('infra-logs-tabpane', require('infrastructures/infra-logs-tabpane.js'));
